requirements.txt
boto3==1.28.57
requests==2.31.0
pytest==7.4.2
moto==4.2.4
Pillow~=10.0.1
setuptools~=65.5.1